+++ app/src/main/java/com/mapbox/maps/testapp/examples/TintFillPatternActivity.kt
@@ -38,9 +38,7 @@ class TintFillPatternActivity : AppCompatActivity() {
 
     binding.mapView.getMapboxMap().loadStyle(
       style(styleUri = Style.MAPBOX_STREETS) {
-        +image(FILL_PATTERN_ID) {
-          bitmap(changeBitmapColor(initialBitmap, randomColor()))
-        }
+        +image(FILL_PATTERN_ID, changeBitmapColor(initialBitmap, randomColor()))
         +fillLayer(FILL_LAYER_ID, STREETS_SOURCE_ID) {
           sourceLayer(SOURCE_LAYER)
           fillPattern(FILL_PATTERN_ID)