+++ app/src/main/java/com/mapbox/maps/testapp/examples/markersandcallouts/ViewAnnotationShowcaseActivity.kt
@@ -103,9 +103,7 @@ class ViewAnnotationShowcaseActivity : AppCompatActivity(), OnMapClickListener, 
   }
 
   private fun prepareStyle(styleUri: String, bitmap: Bitmap) = style(styleUri) {
-    +image(BLUE_ICON_ID) {
-      bitmap(bitmap)
-    }
+    +image(BLUE_ICON_ID, bitmap)
     +geoJsonSource(SOURCE_ID) {
       featureCollection(FeatureCollection.fromFeatures(pointList))
     }