+++ extension-style/src/test/java/com/mapbox/maps/extension/style/image/ImagePluginImplTest.kt
@@ -37,9 +37,7 @@ class ImagePluginImplTest {
   @Test
   fun imageDefaultScaleTest() {
     val image = mockk<Image>(relaxed = true)
-    val imagePlugin = image("imageId") {
-      image(image)
-    }
+    val imagePlugin = image("imageId", image)
     style.addImage(imagePlugin)
     verify { style.addStyleImage("imageId", pixelRatio, image, false, listOf(), listOf(), null) }
   }
@@ -47,8 +45,7 @@ class ImagePluginImplTest {
   @Test
   fun imageUserScaleTest() {
     val image = mockk<Image>(relaxed = true)
-    val imagePlugin = image("imageId") {
-      image(image)
+    val imagePlugin = image("imageId", image) {
       scale(0.4f)
     }
     style.addImage(imagePlugin)
@@ -69,9 +66,7 @@ class ImagePluginImplTest {
       allocatedByteCount = firstArg()
       nativeDataRef
     }
-    val imagePlugin = image("imageId") {
-      bitmap(bitmap)
-    }
+    val imagePlugin = image("imageId", bitmap)
     style.addImage(imagePlugin)
     verify {
       style.addStyleImage(
@@ -104,8 +99,7 @@ class ImagePluginImplTest {
       allocatedByteCount = firstArg()
       nativeDataRef
     }
-    val imagePlugin = image("imageId") {
-      bitmap(bitmap)
+    val imagePlugin = image("imageId", bitmap) {
       scale(1.5f)
     }
     style.addImage(imagePlugin)
@@ -136,9 +130,7 @@ class ImagePluginImplTest {
 
   @Test(expected = RuntimeException::class)
   fun bitmapWrongFormatTest() {
-    val imagePlugin = image("imageId") {
-      bitmap(Bitmap.createBitmap(1, 1, Bitmap.Config.RGB_565))
-    }
+    val imagePlugin = image("imageId", Bitmap.createBitmap(1, 1, Bitmap.Config.RGB_565))
     style.addImage(imagePlugin)
   }
 
@@ -148,8 +140,7 @@ class ImagePluginImplTest {
   @Test
   fun scaleTest() {
     val image = mockk<Image>(relaxed = true)
-    val imagePlugin = image("imageId") {
-      image(image)
+    val imagePlugin = image("imageId", image) {
       scale(1.1f)
     }
     style.addImage(imagePlugin)
@@ -162,8 +153,7 @@ class ImagePluginImplTest {
   @Test
   fun sdfTest() {
     val image = mockk<Image>(relaxed = true)
-    val imagePlugin = image("imageId") {
-      image(image)
+    val imagePlugin = image("imageId", image) {
       sdf(true)
     }
     style.addImage(imagePlugin)
@@ -178,8 +168,7 @@ class ImagePluginImplTest {
   fun stretchXTest() {
     val image = mockk<Image>(relaxed = true)
     val stretchX = listOf(ImageStretches(1.0f, 1.0f), ImageStretches(2.0f, 2.0f))
-    val imagePlugin = image("imageId") {
-      image(image)
+    val imagePlugin = image("imageId", image) {
       stretchX(stretchX)
     }
     style.addImage(imagePlugin)
@@ -194,8 +183,7 @@ class ImagePluginImplTest {
   fun stretchYTest() {
     val image = mockk<Image>(relaxed = true)
     val stretchY = listOf(ImageStretches(1.0f, 1.0f), ImageStretches(2.0f, 2.0f))
-    val imagePlugin = image("imageId") {
-      image(image)
+    val imagePlugin = image("imageId", image) {
       stretchY(stretchY)
     }
     style.addImage(imagePlugin)
@@ -211,8 +199,7 @@ class ImagePluginImplTest {
   fun contentTest() {
     val image = mockk<Image>(relaxed = true)
     val content = ImageContent(1.0f, 2.0f, 3.0f, 4.0f)
-    val imagePlugin = image("imageId") {
-      image(image)
+    val imagePlugin = image("imageId", image) {
       content(content)
     }
     style.addImage(imagePlugin)